Update script to work with XCode 9
XCode 9 requires additional export-ipa.plist values, including a bundle id and provisioning profile setting.
In addition to compatibility with XCode 9, this update removes the need to manually update build.json and export-ipa.plist. They are now template files and will use the values you set in the configuration section of build.sh to replace the template tokens during the build.